Move VolumePluginMgr to standalone file.
